truck: rename admin user service
