+++ src/modules/user/test/user.services.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CreateAdminUserService } from '../services/user-admin.service';
+import { CreateAdminUserService } from '../services/admin-user.service';
 import { UserRepository } from '../repository/user.repository';
 import { AppError } from '../../../common/errors/Error';
 import {
